requirements.txt
tensorflow==2.15
torch==2.3.0 -f https://download.pytorch.org/whl/torch_stable.html
pydicom>=2.4.4
simpleitk>=2.3.1
pillow>=10.3.0
opencv-python>=4.10.0.84
numpy<2
pandas>=2.2.2
nibabel>=5.2.1
torchsummary>=1.5.1
torchvision==0.18.0 -f https://download.pytorch.org/whl/torch_stable.html
scikit-learn>=1.5.0
sklearn-pandas>=2.2.0
seaborn>=0.13.2
plotly>=5.22.0
scikit-image>=0.24.0
scipy>=1.13.1
matplotlib>=3.9.0
ipython>=8.12.3
tqdm>=4.66.4
transformers>=4.42.4
tensorflow-datasets>=4.9.6
vision-transformer-pytorch>=1.0.3
antspyx>=0.5.3
antspynet==0.2.8
monai>=1.3.2
keras-vit>=1.0.1
keras==2.15
itk>=5.4.0
torch-geometric>=2.5.3
torchmetrics>=1.4.1
ipywidgets==8.1.5
notebook==7.2.2
